requirements.txt
cachetools==3.1.1
dataclasses==0.7
fastapi==0.42.0
numpy==1.17.4
pandas==0.25.3
peewee==3.11.2
pydantic==0.32.2
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
starlette==0.12.9
